Initial version
This script manages Rosen Bridge Watcher, running on the Cardano network. It gives 3 options
1. update the docker image
2. just update the heights on the local.yaml and docker compose up -d
3. drop the postgres database and docker compose up -d.

In each case, it queries the tip on ergo and cardano (using BlockFrost) and has you confirm that the values look good before processing. It will decrypt the local.yaml, update with the block heights for Ergo and Cardano, do the operation, then prompt for passwords to re-encrypt the local.yaml.

You'll need to enter your own BlockFrost Project_ID for it to work.