Add files via upload
This script gets the name of the most recently started Docker image, then parses logs counting key phrases that were found in the Haskell files, indicating overall health. The container name changes every run, so had to pull the latest started, as it's most likely to be the SOR, however, if you start an unrelated Docker container just prior to running this script it will not work.